+++ 1.20/src/main/java/io/github/axolotlclient/modules/hud/gui/hud/item/ArmorHud.java
@@ -23,6 +23,7 @@
 package io.github.axolotlclient.modules.hud.gui.hud.item;
 
 import java.util.List;
+import java.util.stream.Stream;
 
 import io.github.axolotlclient.AxolotlClientConfig.api.options.Option;
 import io.github.axolotlclient.AxolotlClientConfig.impl.options.BooleanOption;
@@ -61,16 +62,19 @@ public class ArmorHud extends TextHudEntry {
 	@Override
 	public void renderComponent(GuiGraphics graphics, float delta) {
 		int width = 20;
-		if (showDurabilityNumber.get()) {
-			width += 15;
+		boolean showDurability = showDurabilityNumber.get();
+		boolean showMaxDurability = showMaxDurabilityNumber.get();
+		int labelWidth = showDurability || showMaxDurability ? Stream.concat(Stream.of(client.player.getInventory().getMainHandStack()), client.player.getInventory().armor.stream())
+			.map(stack -> showDurability && showMaxDurability ? (stack.getMaxDamage() - stack.getDamage())+"/"+stack.getMaxDamage() : String.valueOf((showDurability ? stack.getMaxDamage() - stack.getDamage() : stack.getMaxDamage())))
+			.mapToInt(text -> client.textRenderer.getWidth(text)+2).max().orElse(0) : 0;
+		width += labelWidth;
+		if (width != getWidth()) {
+			setWidth(width);
+			onBoundsUpdate();
 		}
-		if (showMaxDurabilityNumber.get()) {
-			width += 15;
-		}
-		setWidth(width);
 		DrawPosition pos = getPos();
 		int lastY = 2 + (4 * 20);
-		renderMainItem(graphics, client.player.getInventory().getMainHandStack(), pos.x() + 2, pos.y() + lastY);
+		renderMainItem(graphics, client.player.getInventory().getMainHandStack(), pos.x() + 2, pos.y() + lastY, labelWidth);
 		lastY = lastY - 20;
 		for (int i = 0; i <= 3; i++) {
 			ItemStack stack = client.player.getInventory().getArmorStack(i).copy();
@@ -86,25 +90,27 @@ public class ArmorHud extends TextHudEntry {
 					}
 				}
 			}
-			renderItem(graphics, stack, pos.x() + 2, lastY + pos.y());
+			renderItem(graphics, stack, pos.x() + 2, lastY + pos.y(), labelWidth);
 			lastY = lastY - 20;
 		}
 	}
 
-	public void renderMainItem(GuiGraphics graphics, ItemStack stack, int x, int y) {
+	public void renderMainItem(GuiGraphics graphics, ItemStack stack, int x, int y, int offset) {
+		renderDurabilityNumber(graphics, stack, x, y);
+		x += offset;
 		String total = String.valueOf(ItemUtil.getTotal(client, stack));
 		if (total.equals("1")) {
 			total = null;
 		}
 		graphics.drawItem(stack, x, y);
 		graphics.drawItemInSlot(client.textRenderer, stack, x, y, total);
-		renderDurabilityNumber(graphics, stack, x, y);
 	}
 
-	public void renderItem(GuiGraphics graphics, ItemStack stack, int x, int y) {
+	public void renderItem(GuiGraphics graphics, ItemStack stack, int x, int y, int offset) {
+		renderDurabilityNumber(graphics, stack, x, y);
+		x += offset;
 		graphics.drawItem(stack, x, y);
 		graphics.drawItemInSlot(client.textRenderer, stack, x, y);
-		renderDurabilityNumber(graphics, stack, x, y);
 	}
 
 	private void renderDurabilityNumber(GuiGraphics graphics, ItemStack stack, int x, int y) {
@@ -121,13 +127,24 @@ public class ArmorHud extends TextHudEntry {
 
 	@Override
 	public void renderPlaceholderComponent(GuiGraphics graphics, float delta) {
+		int width = 20;
+		boolean showDurability = showDurabilityNumber.get();
+		boolean showMaxDurability = showMaxDurabilityNumber.get();
+		int labelWidth = showDurability || showMaxDurability ? Stream.concat(Stream.of(client.player.getInventory().getMainHandStack()), client.player.getInventory().armor.stream())
+			.map(stack -> showDurability && showMaxDurability ? (stack.getMaxDamage() - stack.getDamage())+"/"+stack.getMaxDamage() : String.valueOf((showDurability ? stack.getMaxDamage() - stack.getDamage() : stack.getMaxDamage())))
+			.mapToInt(text -> client.textRenderer.getWidth(text)+2).max().orElse(0) : 0;
+		width += labelWidth;
+		if (width != getWidth()) {
+			setWidth(width);
+			onBoundsUpdate();
+		}
 		DrawPosition pos = getPos();
 		int lastY = 2 + (4 * 20);
-		renderItem(graphics, placeholderStacks[4], pos.x() + 2, pos.y() + lastY);
+		renderItem(graphics, placeholderStacks[4], pos.x() + 2, pos.y() + lastY, labelWidth);
 		lastY = lastY - 20;
 		for (int i = 0; i <= 3; i++) {
 			ItemStack item = placeholderStacks[i];
-			renderItem(graphics, item, pos.x() + 2, lastY + pos.y());
+			renderItem(graphics, item, pos.x() + 2, lastY + pos.y(), labelWidth);
 			lastY = lastY - 20;
 		}
 	}