+++ 1.21/src/main/java/io/github/axolotlclient/modules/freelook/Freelook.java
@@ -22,6 +22,9 @@
 
 package io.github.axolotlclient.modules.freelook;
 
+import java.util.ArrayDeque;
+import java.util.Deque;
+
 import com.mojang.blaze3d.platform.InputUtil;
 import io.github.axolotlclient.AxolotlClient;
 import io.github.axolotlclient.AxolotlClientConfig.api.options.OptionCategory;
@@ -39,71 +42,74 @@ import net.minecraft.entity.Entity;
 
 public class Freelook extends AbstractModule {
 
-	private static final Freelook Instance = new Freelook();
-	private static KeyBind KEY = KeyBinds.getInstance().register(new KeyBind("key.freelook", InputUtil.KEY_V_CODE, "category.axolotlclient"));
+	private static final Freelook INSTANCE = new Freelook();
+	private static final KeyBind KEY = KeyBinds.getInstance().register(new KeyBind("key.freelook", InputUtil.KEY_V_CODE, "category.axolotlclient"));
+	private static final KeyBind KEY_ALT = KeyBinds.getInstance().register(new KeyBind("key.freelook.alt", InputUtil.UNKNOWN_KEY.getKeyCode(), "category.axolotlclient"));
 	public final ForceableBooleanOption enabled = new ForceableBooleanOption("enabled", false);
 	private final MinecraftClient client = MinecraftClient.getInstance();
 	private final OptionCategory category = OptionCategory.create("freelook");
 	private final StringArrayOption mode = new StringArrayOption("mode",
 		new String[]{"snap_perspective", "freelook"},
 		"freelook", value -> FeatureDisabler.update());
+	private final BooleanOption invert = new BooleanOption("invert", false);
 	private final EnumOption<Perspective> perspective = new EnumOption<Perspective>("perspective", Perspective.class,
 		Perspective.THIRD_PERSON_BACK);
-	private final BooleanOption invert = new BooleanOption("invert", false);
 	private final BooleanOption toggle = new BooleanOption("toggle", false);
-	public boolean active;
+	private final EnumOption<Perspective> perspectiveAlt = new EnumOption<>("perspective.alt", Perspective.class,
+		Perspective.THIRD_PERSON_FRONT);
+	private final BooleanOption toggleAlt = new BooleanOption("toggle.alt", false);
+	private final WrappedValue active = new WrappedValue(), activeAlt = new WrappedValue();
 	private float yaw, pitch;
-	private Perspective previousPerspective;
-
-	public Freelook() {
-	}
+	private final Deque<Perspective> previousPerspectives = new ArrayDeque<>();
 
 	public static Freelook getInstance() {
-		return Instance;
+		return INSTANCE;
 	}
 
 	@Override
 	public void init() {
 		category.add(enabled, mode, perspective, invert, toggle);
+		category.add(perspectiveAlt, toggleAlt);
 		AxolotlClient.CONFIG.addCategory(category);
 	}
 
 	@Override
 	public void tick() {
-		if (!enabled.get())
-			return;
+		if (!enabled.get() || client.currentScreen != null) return;
+		tickSet(toggle, KEY, perspective, active);
+		tickSet(toggleAlt, KEY_ALT, perspectiveAlt, activeAlt);
+	}
 
+	private void tickSet(BooleanOption toggle, KeyBind key, EnumOption<Perspective> perspective, WrappedValue active) {
 		if (toggle.get()) {
-			if (KEY.wasPressed()) {
-				if (active) {
-					stop();
+			if (key.wasPressed()) {
+				if (active.val) {
+					stop(active);
 				} else {
-					start();
+					start(perspective.get(), active);
 				}
 			}
 		} else {
-			if (KEY.isPressed()) {
-				if (!active) {
-					start();
+			if (key.isPressed()) {
+				if (!active.val) {
+					start(perspective.get(), active);
 				}
-			} else if (active) {
-				stop();
+			} else if (active.val) {
+				stop(active);
 			}
 		}
 	}
 
-	private void stop() {
-		active = false;
+	private void stop(WrappedValue active) {
+		active.val = false;
 		client.worldRenderer.scheduleTerrainUpdate();
-		setPerspective(previousPerspective);
+		setPerspective(previousPerspectives.pop());
 	}
 
-	private void start() {
-		active = true;
-
-
-		previousPerspective = client.options.getPerspective();
-		setPerspective(perspective.get());
+	private void start(Perspective perspective, WrappedValue active) {
+		previousPerspectives.push(client.options.getPerspective());
+		active.val = true;
+		setPerspective(perspective);
 
 		Entity camera = client.getCameraEntity();
 
@@ -121,7 +127,7 @@ public class Freelook extends AbstractModule {
 	}
 
 	public boolean consumeRotation(double dx, double dy) {
-		if (!active || !enabled.get() || !mode.get().equals("freelook"))
+		if (!(active.val || activeAlt.val) || !enabled.get() || !mode.get().equals("freelook"))
 			return false;
 
 		if (!invert.get())
@@ -131,8 +137,8 @@ public class Freelook extends AbstractModule {
 			|| MinecraftClient.getInstance().options.getPerspective().isFirstPerson())
 			dy *= -1;
 
-		yaw += dx * 0.15F;
-		pitch += dy * 0.15F;
+		yaw += (float) (dx * 0.15F);
+		pitch += (float) (dy * 0.15F);
 
 		if (pitch > 90) {
 			pitch = 90;
@@ -145,14 +151,14 @@ public class Freelook extends AbstractModule {
 	}
 
 	public float yaw(float defaultValue) {
-		if (!active || !enabled.get() || !mode.get().equals("freelook"))
+		if (!(active.val || activeAlt.val) || !enabled.get() || !mode.get().equals("freelook"))
 			return defaultValue;
 
 		return yaw;
 	}
 
 	public float pitch(float defaultValue) {
-		if (!active || !enabled.get() || !mode.get().equals("freelook"))
+		if (!(active.val || activeAlt.val) || !enabled.get() || !mode.get().equals("freelook"))
 			return defaultValue;
 
 		return pitch;
@@ -160,5 +166,13 @@ public class Freelook extends AbstractModule {
 
 	public boolean needsDisabling() {
 		return mode.get().equals("freelook");
+	}
+
+	public boolean isActive(){
+		return active.val || activeAlt.val;
+	}
+
+	private static class WrappedValue {
+		boolean val;
 	}
 }