+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - Fix various bugs & crashes
 - Skip Celestial sky packs while scanning for custom skies
 - (1.8.9, 1.16) Enable texture nine-slicing for all ButtonWidgets
+- Update bedwars module
+- add secondary keybinding for freelook
 
 ### 3.1.0